requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.7
django-crispy-forms==2.1
django-filter==23.4
sqlparse==0.4.4
tzdata==2023.3
